pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codeowners-coverage"
version = "0.1.0"
description = "Measure and enforce CODEOWNERS coverage"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Apache-2.0" }
authors = [
    { name = "Sentry Team", email = "hello@sentry.io" }
]
dependencies = [
    "pathspec>=0.12.0",  # GitIgnore pattern matching
    "PyYAML>=6.0",       # Config file parsing
    "click>=8.0",        # CLI framework
    "requests>=2.31.0",  # For GitHub API
    "ollama>=0.3.0",     # For local LLM integration
]

[project.scripts]
codeowners-coverage = "codeowners_coverage.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
